--- board/atmel/sama5d2_ptc/Makefile
@@ -1,8 +1,0 @@
-#
-# Copyright (C) 2016 Atmel
-#		     Wenyou Yang <wenyou.yang@atmel.com>
-#
-# SPDX-License-Identifier:	GPL-2.0+
-#
-
-obj-y += sama5d2_ptc.o
