+++ board/atmel/sama5d2_ptc_ek/Kconfig
@@ -1,7 +1,7 @@
-if TARGET_SAMA5D2_PTC
+if TARGET_SAMA5D2_PTC_EK
 
 config SYS_BOARD
-	default "sama5d2_ptc"
+	default "sama5d2_ptc_ek"
 
 config SYS_VENDOR
 	default "atmel"
@@ -10,6 +10,6 @@ config SYS_SOC
 	default "at91"
 
 config SYS_CONFIG_NAME
-	default "sama5d2_ptc"
+	default "sama5d2_ptc_ek"
 
 endif
